// webApp/src/app/enums/socket-events.enums.ts
export enum SocketEvents {
  Join = 'joinRoom',
  Message = 'message',
  Login = 'login',
  ListTrigger = 'listTrigger',
  RoomCreated = 'roomCreated',
  OpponentListener = 'opponentListener',
  LeaveRoom = 'leaveRoom',
  LeaveRoomButton = 'leaveRoomByClick',
  LetsPlay = 'letsPlay',
  RandomNumber = 'randomNumber',
  SendNumber = 'sendNumber',
  ActivateTurn = 'activateYourTurn',
  IsNotDivisible = 'isNotDivisible',
  OnReady = 'onReady',
  DeActivateJoining = 'deActivateJoin',
  GameOver = 'gameOver',
}